EL [feature/all-virtual-players -> dev] WIP all-virtual-players merge

See merge request polytechnique-montr-al/log2990/20213/equipe-304/log2990-304!23

+++ client/src/app/services/game.service.ts
@@ -193,8 +193,10 @@ export class GameService {
                 }
                 const lettersToAdd: ScrabbleLetter[] = this.game.stock.exchangeLetters(lettersToRemove);
                 for (let i = 0; i < lettersToAdd.length; i++) {
-                    this.rackService.removeLetter(lettersToRemove[i]);
-                    this.rackService.addLetter(lettersToAdd[i]);
+                    if (!(player instanceof VirtualPlayer)) {
+                        this.rackService.removeLetter(lettersToRemove[i]);
+                        this.rackService.addLetter(lettersToAdd[i]);
+                    }
                     player.letters.push(lettersToAdd[i]);
                 }
                 if (this.game.gameMode === GameType.MultiPlayer) {
